More cleanup code Converted some missing foreach+if loops into Linq syntax Changed loop count variables to .Count in Linq Commented some not used variables Added some StringComparison.InvariantCultureIgnoreCase to EndsWith/StartsWith (note: not necessary for all but keep on for a while) Removed some unnecessary elses Other minor changes

+++ src/L2dotNET.GameService/templates/NpcTemplate.cs
@@ -58,7 +58,7 @@ namespace L2dotNET.GameService.Templates
             Name = set.GetString("name");
             Title = set.GetString("title", "");
             CantBeChampionMonster = Title.Equals("Quest Monster", StringComparison.InvariantCultureIgnoreCase);
-            Level = set.GetByte("level", (byte)1);
+            Level = set.GetByte("level", 1);
             Exp = set.GetInt("exp", 0);
             Sp = set.GetInt("sp", 0);
             RHand = set.GetInt("rHand", 0);
